feat: Player Animations and some assets
I changed some things in PlayerLocomotion,PlayerGrab and PlayerMachete to allow for the animations to play properly. I also changed PlayerLocomotion to rotate the character in the direction it is walking towards.

I also added a lot of assets, some of them are not fully added so keep that in mind.