fix: added onto demo scene
i don't know where the tap is meant to go so i put it where it felt right.
be mindful of the public variables you have to place, in animation objects is 3 objects in the tap with Animation in their name and for the buckets is the Tap Bucket in the tap and Hand Bucket in jeffrey.